Remove analyze folder
